Feat: Complete wiring of drains history table
