+++ src/components/v5/frame/ColonyHome/partials/TokenBalance/TokenBalance.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { useLocation } from 'react-router-dom';
 
+import { currencySymbolMap } from '~constants/currency.ts';
 import { useCurrencyContext } from '~context/CurrencyContext/CurrencyContext.ts';
 import { COLONY_BALANCES_ROUTE } from '~routes/index.ts';
 import Numeral from '~shared/Numeral/index.ts';
-import { currencySymbolMap } from '~utils/currency/config.ts';
 import { formatText } from '~utils/intl.ts';
 import WidgetBox from '~v5/common/WidgetBox/index.ts';
